src/streaming/streaming.ts
import { StreamState, StreamReader, StreamWriter } from '../../spec/Streaming';

export function createStreamState(): StreamState {
  return {
    mode: 'read',
    options: {
      bufferSize: 8192, // 8KB default buffer
      cacheSize: 1024 * 1024, // 1MB cache
      memoryLimit: 1024 * 1024 * 1024, // 1GB memory limit
      compression: true,
      async: true,
      parallel: true
    },
    status: {
      position: 0,
      rowCount: 0,
      cellCount: 0,
      memoryUsed: 0,
      isOpen: false
    },
    cache: {
      strings: new Map(),
      styles: new Map(),
      formulas: new Map()
    }
  };
}

export function createStreamReader(bufferSize: number = 8192): StreamReader {
  return {
    position: 0,
    bufferSize,
    async read(options = {}) {
      const {
        rows = 100,
        cells = 1000,
        parseOptions = {
          values: true,
          formulas: true,
          styles: false
        }
      } = options;

      // Placeholder for actual streaming implementation
      // In a real implementation, this would read from a file or stream
      return {
        data: [],
        position: this.position,
        hasMore: false
      };
    }
  };
}

export function createStreamWriter(bufferSize: number = 8192): StreamWriter {
  return {
    position: 0,
    bufferSize,
    async write(data, options = {}) {
      const {
        flush = false,
        styles = false,
        formulas = true
      } = options;

      // Placeholder for actual streaming implementation
      // In a real implementation, this would write to a file or stream
    },
    async flush() {
      // Placeholder for actual flush implementation
      // In a real implementation, this would flush the buffer to disk/stream
    }
  };
}

export function setStreamOptions(state: StreamState, options: Partial<StreamState['options']>): StreamState {
  return {
    ...state,
    options: {
      ...state.options,
      ...options
    }
  };
}

export function updateStreamStatus(state: StreamState, status: Partial<StreamState['status']>): StreamState {
  return {
    ...state,
    status: {
      ...state.status,
      ...status
    }
  };
}

export function clearStreamCache(state: StreamState): StreamState {
  return {
    ...state,
    cache: {
      strings: new Map(),
      styles: new Map(),
      formulas: new Map()
    },
    status: {
      ...state.status,
      memoryUsed: 0
    }
  };
}

export function addToCacheString(state: StreamState, index: number, value: string): StreamState {
  const newStrings = new Map(state.cache.strings);
  newStrings.set(index, value);
  
  return {
    ...state,
    cache: {
      ...state.cache,
      strings: newStrings
    }
  };
}

export function addToCacheStyle(state: StreamState, index: number, style: any): StreamState {
  const newStyles = new Map(state.cache.styles);
  newStyles.set(index, style);
  
  return {
    ...state,
    cache: {
      ...state.cache,
      styles: newStyles
    }
  };
}

export function addToCacheFormula(state: StreamState, cell: string, formula: string): StreamState {
  const newFormulas = new Map(state.cache.formulas);
  newFormulas.set(cell, formula);
  
  return {
    ...state,
    cache: {
      ...state.cache,
      formulas: newFormulas
    }
  };
}

export function isMemoryWithinLimit(state: StreamState, additionalBytes: number): boolean {
  return (state.status.memoryUsed + additionalBytes) <= state.options.memoryLimit;
}

export function calculateMemoryUsage(state: StreamState): number {
  // Rough estimation of memory usage
  const stringsSize = Array.from(state.cache.strings.values())
    .reduce((total, str) => total + str.length * 2, 0);
  
  const stylesSize = state.cache.styles.size * 100; // Rough estimate of 100 bytes per style
  const formulasSize = Array.from(state.cache.formulas.values())
    .reduce((total, formula) => total + formula.length * 2, 0);
  
  return stringsSize + stylesSize + formulasSize;
}